Add ability to log directly to journald
Iff the `conf.log` setting is set to `'journald'` then attempt to
load the journald Winston transport. If that fails (for example,
because we're not running on Linux or we're running on a system
that does not use systemd or does not use journald) and kiwi is
called wih the `-f` option, then it will fall back to logging to a
file called `kiwi.log`. Otherwise, logging will continue to work
in the same way it does now.

Setting `process.title` has also been moved from the bottom of the
file to the top, so that any initial logging messages are recorded
(by journald/syslog) using the process' final name.